requirements.txt
aiofiles==23.1.0
calc==1.0
websocket-client==1.5.2
websockets==11.0.3
